[#172382500] Move modals from root to 'screens/modal'
